+++ socket-server/namespaces/api-server.ts
@@ -1,5 +1,5 @@
 import socketIO from "socket.io";
-import controller from "../controllers";
+import { itemController } from "../controllers";
 
 const getApiServerNamespace = (io: socketIO.Server) => {
   const apiServerNamespace = io.of("/api-server");
@@ -7,19 +7,24 @@ const getApiServerNamespace = (io: socketIO.Server) => {
   apiServerNamespace.on("connection", async (socket: socketIO.Socket) => {
     socket.on("disconnecting", async () => {});
 
-    socket.on("cancelBooking", async (scheduleId: string, seatIdArray: [string]) => {
-      await controller.setUnSoldSeats(socket.id, scheduleId, seatIdArray);
-      const seats = await controller.getSeatDataByScheduleId(scheduleId);
-      const counts = await controller.getAllClassCount(scheduleId);
+    socket.on(
+      "cancelBooking",
+      async (userId: string, scheduleId: string, seatIdArray: [string]) => {
+        await itemController.setUnSoldSeats(userId, scheduleId, seatIdArray);
 
-      io.of("/client").to(`${scheduleId}-booking`).emit("receiveSeat", seats);
-      io.of("/client").to(`${scheduleId}-booking`).emit("receiveCount", counts);
-      io.of("/client").to(`${scheduleId}-count`).emit("receiveCount", counts);
-    });
+        const seats = await itemController.getSeatDataByScheduleId(scheduleId);
+        const counts = await itemController.getAllClassCount(scheduleId);
 
-    socket.on("bookSeat", async (scheduleId: string, seatIdArray: [string]) => {
-      await controller.setSoldSeats(socket.id, scheduleId, seatIdArray);
-      const seats = await controller.getSeatDataByScheduleId(scheduleId);
+        io.of("/client").to(`${scheduleId}-booking`).emit("receiveSeat", seats);
+        io.of("/client").to(`${scheduleId}-booking`).emit("receiveCount", counts);
+        io.of("/client").to(`${scheduleId}-count`).emit("receiveCount", counts);
+      },
+    );
+
+    socket.on("bookSeat", async (userId: string, scheduleId: string, seatIdArray: [string]) => {
+      await itemController.setSoldSeats(userId, scheduleId, seatIdArray);
+
+      const seats = await itemController.getSeatDataByScheduleId(scheduleId);
 
       io.of("/client").to(`${scheduleId}-booking`).emit("receiveSeat", seats);
     });
