+++ socket-server/utils/index.ts
@@ -1,0 +1,7 @@
+const getKey = (id: string, key: string) => {
+  return `${id}:${key}`;
+};
+
+const getExpireKey = () => {};
+
+export { getKey, getExpireKey };
